+++ src/core/thread/qmutex_win.cpp
@@ -70,7 +70,7 @@ void QMutex::unlock()
 {
     const int none = 0;
 
-    Q_ASSERT_X(d->owner == GetCurrentThreadId(), "QMutex::unlock()",
+    Q_ASSERT_X(static_cast<DWORD>(d->owner) == GetCurrentThreadId(), "QMutex::unlock()",
                "A mutex must be unlocked in the same thread that locked it.");
 
     if (!--d->count) {
